fixed string coorsion of null
