+++ packages/core/components/Filters/index.ts
@@ -1,3 +1,5 @@
 export { default } from './Filters'
 
-export { filterOrderOptions, filterStyleOptions, handleSorting, useFilters } from './Filters'
+export { filterOrderOptions, filterStyleOptions, useFilters } from './Filters'
+
+export { handleSorting } from './helpers/handleSorting'
